refactor: move book_members_live
