--- apps/app_web/lib/app_web/live/book_member_live/index.html.heex
@@ -1,17 +1,0 @@
-<div class="max-w-prose mx-auto">
-  <.tile
-    :if={Rights.can_member_invite_new_member?(@current_member)}
-    navigate={~p"/books/#{@book}/invite"}
-  >
-    <.icon name="person-add" />
-    <%= gettext("Invite a new member") %>
-  </.tile>
-
-  <.member_tile :for={member <- @linked_members} member={member} />
-
-  <.heading :if={not Enum.empty?(@independent_members)} level={:section} class="mt-6">
-    <%= gettext("Pending members") %>
-  </.heading>
-
-  <.member_tile :for={member <- @independent_members} member={member} />
-</div>
