+++ apps/app_web/lib/app_web/router.ex
@@ -75,7 +75,7 @@ defmodule AppWeb.Router do
       live "/books/:book_id/edit", BookLive.Form, :edit
 
       live "/books/:book_id/invite", BookInvitationsLive, :show
-      live "/books/:book_id/members", BookMemberLive.Index, :index
+      live "/books/:book_id/members", BookMemberLive, :index
       live "/books/:book_id/transfers", MoneyTransferLive.Index, :index
       live "/books/:book_id/transfers/new", MoneyTransferLive.Form, :new
       live "/books/:book_id/transfers/:money_transfer_id/edit", MoneyTransferLive.Form, :edit
